Add short wait to page focus for 'open_link'.
Occasionally, if the new tab is not yet loaded, it would crash the
original page.

This should probably be done in a loop and if the page is still not
focus-able after perhaps 2 seconds, then just don't focus. The way it is
now will still crash the page if it isn't available after 100ms.